requirements-dev.txt
black==20.8b1
django-debug-toolbar==2.1
ipython==7.11.1
jupyter==1.0.0
mypy==0.761
mypy-extensions==0.4.3
notebook==6.0.2
pre-commit==1.21.0
pylint==2.4.4
pylint-django==2.3.0
